requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beanie==2.1.0
click==8.3.2
dnspython==2.8.0
fastapi==0.135.3
fastapi-utils==0.8.0
h11==0.16.0
idna==3.11
lazy-model==0.4.0
psutil==5.9.8
pydantic==2.13.0
pydantic_core==2.46.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
